+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Changelog
 unreleased
 ==========
 * feat: Compliance number filter added
+
+1.1.0 (2022-06-15)
+==================
+* feat: Change view title and icon tooltips name changed
 * feat: Added compliance number field
 
 1.0.0 (2022-04-25)